internal socket class
